added entites

+++ src/main/java/org/example/grad_gateway2/DTO/RegisterDTO.java
@@ -27,5 +27,5 @@ public class RegisterDTO {
     @NotBlank(message = "Role is required")
     private String role;
 
-    private int companyId;
+    private long companyId;
 }
